Added SWBuf::stripPrefix method which works nice 		for returning and stripping stuff like 		'lemma' from "lemma:G1234" 	Added part params to XMLTag::setAttribute 	Changed OSISStrongs to only strip lemma parts with 		strong prefix 	Applied fix for newer version of gcc which errors if 		create perms are not passed to ::open 		(Deji Akingunola <dakingun at gmail dot com>)
git-svn-id: https://crosswire.org/svn/sword/trunk@2096 bcd7d363-81e1-0310-97ec-a550e20fc99c

+++ src/modules/filters/utf8transliterator.cpp
@@ -349,7 +349,7 @@ char UTF8Transliterator::processText(SWBuf &text, const SWKey *key, const SWModu
                         scripts[i] = false;
                 }
 
-                for (i = 0; i < len; i++) {
+                for (i = 0; i < (unsigned long)len; i++) {
                         j = ublock_getCode(source[i]);
 			scripts[SE_LATIN] = true;
 			switch (j) {
